tiff package refactor
