+++ src/main/java/mil/nga/tiff/FieldType.java
@@ -1,4 +1,4 @@
-package mil.nga.tiff.io;
+package mil.nga.tiff;
 
 /**
  * Field Types
