+++ src/main/java/mil/nga/tiff/FileDirectoryEntry.java
@@ -1,4 +1,4 @@
-package mil.nga.tiff.io;
+package mil.nga.tiff;
 
 /**
  * TIFF File Directory Entry
