+++ src/main/java/mil/nga/tiff/Rasters.java
@@ -1,4 +1,4 @@
-package mil.nga.tiff.io;
+package mil.nga.tiff;
 
 import mil.nga.tiff.util.TiffException;
