+++ src/main/java/mil/nga/tiff/TempTest.java
@@ -1,4 +1,4 @@
-package mil.nga.tiff.io;
+package mil.nga.tiff;
 
 import java.io.File;
